Revert "Fatosynew"
