+++ package.json
@@ -35,12 +35,8 @@
     "css-loader": "^6.4.0",
     "css-minimizer-webpack-plugin": "^4.1.0",
     "dayjs": "^1.11.5",
-    "file-saver": "^2.0.5",
     "html-webpack-plugin": "^5.3.2",
     "jquery": "^3.6.0",
-    "js-table2excel": "^1.0.3",
-    "jsdoc": "^3.6.11",
-    "jszip": "^3.10.1",
     "mini-css-extract-plugin": "^2.6.1",
     "mitt": "^3.0.0",
     "sass": "^1.54.3",
