+++ public/manifest.json
@@ -6,8 +6,7 @@
   "background": { 
     "service_worker": "background.js" 
   },
-  "permissions": ["storage", "activeTab", "tabs", "contextMenus", "scripting", "downloads", "cookies"],
-  "host_permissions":["https://*.jd.com/*"],
+  "permissions": ["storage", "activeTab", "tabs", "contextMenus", "scripting", "downloads"],
   "action": {
     "default_title": "popup.html",
     "default_icon": "logo.png",
@@ -16,19 +15,16 @@
   "icons": { "16": "logo.png", "48": "logo.png", "128": "logo.png" },
   "content_scripts": [
     {
-      "matches": ["https://*.1688.com/*", "https://*.tmall.com/*", "https://*.jd.com/*",
+      "matches": ["https://*.1688.com/*", "https://*.tmall.com/*", "https://*.jd.com/*", 
       "https://*.taobao.com/*", "https://*.alibaba.com/*", "https://*.yangkeduo.com/*",
       "https://*.pingduoduo.com/*", "http://*.junchenlun.com/*", "https://www.google.com/"
-      ],
+    ],
       "js": ["content.js"]
     }
   ],
   "web_accessible_resources": [{
-    "resources": ["inject.js","content.js", "content.js.map", "inject.js.map", "jdi.js"],
+    "resources": ["inject.js","content.js", "content.js.map", "inject.js.map"],
     "matches": ["<all_urls>"]
-  }],
-  "content_security_policy": {
-    "script-src": "unsafe-inline"
-  }
+  }]
 }
