+++ src/api/contentApi/dayjs.js
@@ -6,25 +6,4 @@
 // const dayjs = require('dayjs')
 import dayjs from "dayjs"
 
-class zTime{
-    constructor(){
-        this.nowtime = dayjs()
-    }
-
-    ymdhms(){
-        return dayjs().format('YYYY-MM-DD HH:mm:ss')
-    }
-
-    ymd(){
-        return dayjs().format('YYYY-MM-DD')
-    }
-
-    ymd2(){
-        return dayjs().format('YYYYMMDD')
-    }
-
-}
-
-let ztime = new zTime()
-
-export default {dayjs: dayjs(), ztime}
+export default {dayjs: dayjs()}
