+++ src/api/contentApi/storage.js
@@ -4,7 +4,6 @@
  * @LastEditTime: 2022-10-28 15:16:53
  */
 const Storage = {
-
     set(obj){
         return new Promise( (resolve, reject) => {
