--- src/pages/jd/js/JDinject.js
@@ -1,1 +1,0 @@
-console.log('dasdasdasfdaf')
