--- src/pages/jd/js/diagnosis.js
@@ -1,4 +1,0 @@
-
-
-
-
