+++ src/pages/tb/app.vue
@@ -186,6 +186,11 @@
 
 <script setup>
 
+// let aa = API.dayjs.format('YYYY-MM-DD HH:mm:ss')
+// console.log('aa: ', aa);
+const userstore = userStore()
+const { location } = storeToRefs(userstore)
+
 //---------------单纯字符串变量不可使用reactive---------
 //-----ref定义的数据：操作数据需要.value，读取数据时模板中直接读取不需要
 let currentHref = ref('')
@@ -490,7 +495,6 @@ const onDragstop = (e) => {
       })
     getUserInfo()
    })
-
 
 </script>
 <style lang="scss" scoped>
